// packages/discord/src/operations/guildSearch.ts
export const GuildSearchInput = /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
  guild_id: Schema.String.pipe(T.PathParam()),
  sort_by: Schema.optional(Schema.String),
  sort_order: Schema.optional(Schema.String),
  content: Schema.optional(Schema.String),
  slop: Schema.optional(Schema.Number),
  author_id: Schema.optional(Schema.String),
  author_type: Schema.optional(Schema.String),
  mentions: Schema.optional(Schema.String),
  mentions_role_id: Schema.optional(Schema.String),
  replied_to_user_id: Schema.optional(Schema.String),
  replied_to_message_id: Schema.optional(Schema.String),
  mention_everyone: Schema.optional(Schema.Boolean),
  min_id: Schema.optional(Schema.String),
  max_id: Schema.optional(Schema.String),
  limit: Schema.optional(Schema.Number),
  offset: Schema.optional(Schema.Number),
  has: Schema.optional(Schema.String),
  link_hostname: Schema.optional(Schema.String),
  embed_provider: Schema.optional(Schema.String),
  embed_type: Schema.optional(Schema.String),
  attachment_extension: Schema.optional(Schema.String),
  attachment_filename: Schema.optional(Schema.String),
  pinned: Schema.optional(Schema.Boolean),
  include_nsfw: Schema.optional(Schema.Boolean),
  channel_id: Schema.optional(Schema.String),
}).pipe(T.Http({ method: "GET", path: "/guilds/{guild_id}/messages/search" }));
export type GuildSearchInput = typeof GuildSearchInput.Type;
// ...
export type GuildSearchOutput = typeof GuildSearchOutput.Type;

// The operation
export const guildSearch = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: GuildSearchInput,
  outputSchema: GuildSearchOutput,
  errors: [BadRequest, Forbidden, NotFound] as const,
}));